UI Module: Move CellController to its own file
