+++ EssentialFeed/EssentialFeediOS/Shared UI/ViewControllers/ListViewController.swift
@@ -5,24 +5,6 @@
 import UIKit
 import EssentialFeed
 
-public struct CellController {
-	let dataSource: UITableViewDataSource
-	let delegate: UITableViewDelegate?
-	let dataSourcePrefetching: UITableViewDataSourcePrefetching?
-
-	public init(_ dataSource: UITableViewDataSource & UITableViewDelegate & UITableViewDataSourcePrefetching) {
-		self.dataSource = dataSource
-		self.delegate = dataSource
-		self.dataSourcePrefetching = dataSource
-	}
-
-	public init(_ dataSource: UITableViewDataSource) {
-		self.dataSource = dataSource
-		self.delegate = nil
-		self.dataSourcePrefetching = nil
-	}
-}
-
 public final class ListViewController: UITableViewController, UITableViewDataSourcePrefetching, ResourceLoadingView, ResourceErrorView {
 	@IBOutlet private(set) public var errorView: ErrorView?
